VoteAdmin/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
pymongo==3.11.4
Flask-Bcrypt==0.7.1
gunicorn==20.1.0
Werkzeug==2.1.1
Flask-PyMongo==2.3.0
